Forms and resources final touches
