+++ package.json
@@ -9,7 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@heroicons/react": "^2.1.5",
+    "@nextui-org/breadcrumbs": "^2.0.13",
     "@nextui-org/card": "^2.0.34",
+    "@nextui-org/date-picker": "^2.1.8",
     "@nextui-org/input": "^2.2.5",
     "@nextui-org/react": "^2.4.8",
     "@nextui-org/system": "^2.2.6",
@@ -24,9 +27,9 @@
     "flowbite": "^2.5.2",
     "flowbite-react": "^0.10.2",
     "formidable": "^3.5.1",
-    "framer-motion": "^11.2.4",
-    "multer": "^1.4.5-lts.1",
+    "framer-motion": "^11.11.11",
     "lucide-react": "^0.453.0",
+    "multer": "^1.4.5-lts.1",
     "net": "^1.0.2",
     "next": "^14.2.3",
     "next-connect": "^1.0.0",
@@ -42,7 +45,7 @@
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.16",
     "@types/papaparse": "^5.3.14",
-    "@types/react": "^18",
+    "@types/react": "^18.3.12",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
