+++ src/pages/contact/index.tsx
@@ -7,6 +7,8 @@ import Link from 'next/link';
 import { DatePicker } from "@nextui-org/date-picker";
 import { Card, CardBody } from '@nextui-org/react';
 import { DateValue, today } from '@internationalized/date';
+import { MapPinIcon, PhoneIcon } from '@heroicons/react/24/outline';
+import { FaCheckCircle, FaExclamationTriangle, FaPaperPlane } from 'react-icons/fa';
 
 const ContactUs = () => {
   const [formData, setFormData] = useState({
@@ -21,7 +23,8 @@ const ContactUs = () => {
   const [isLoading, setIsLoading] = useState(false);
   const [modalContent, setModalContent] = useState({
     title: "",
-    message: ""
+    message: "",
+    type: ""
   });
 
   const handleChange = (e: { target: { name: any; value: any; }; }) => {
@@ -38,8 +41,9 @@ const ContactUs = () => {
 
     if (!formDataToSend.fullName || !formDataToSend.date || !formDataToSend.email || !formDataToSend.comment) {
       setModalContent({
-        title: "Validation Error",
-        message: "Please fill in all required fields."
+        title: "Oops! Missing Information",
+        message: "Looks like some fields are feeling lonely! 📝 Please fill in all the required ones.",
+        type: "error"
       });
       onOpen();
       return;
@@ -48,8 +52,9 @@ const ContactUs = () => {
     setIsLoading(true);
     onOpen();
     setModalContent({
-      title: "Sending your message",
-      message: "Please hold on while we process your request..."
+      title: "Message in Transit",
+      message: "Your message is flying through the digital clouds! ✈️",
+      type: "loading"
     });
 
     try {
@@ -62,20 +67,23 @@ const ContactUs = () => {
       const result = await response.json();
       if (response.ok) {
         setModalContent({
-          title: "Message Sent Successfully!",
-          message: "Thank you for contacting us. We'll get back to you within 24-48 business hours."
+          title: "Message Delivered Successfully!",
+          message: "Your message has landed safely in our inbox! 🎯 We'll get back to you faster than a caffeinated scientist!",
+          type: "success"
         });
         setFormData({ fullName: "", date: today('UTC') as DateValue, email: "", phone: "", comment: "" });
       } else {
         setModalContent({
-          title: "Error",
-          message: `Error: ${result.message}`
+          title: "Message Delivery Failed",
+          message: `Whoops! Our digital carrier pigeon got lost! 🐦 Error: ${result.message}`,
+          type: "error"
         });
       }
     } catch (error) {
       setModalContent({
-        title: "Error",
-        message: "There was an error sending your message. Please try again later."
+        title: "Communication Breakdown",
+        message: "Houston, we have a problem! 🚀 Our message system needs a coffee break. Please try again later!",
+        type: "error"
       });
     } finally {
       setIsLoading(false);
@@ -96,8 +104,23 @@ const ContactUs = () => {
               Contact Us
             </h1>
             <p className="mt-5 text-left text-gray-600 max-w-sm">
-              Have a question? We would love to help! Fill out this form and we’ll get back to you as soon as possible.
+              Have a question? We would love to help! Fill out this form and we'll get back to you as soon as possible.
             </p>
+            
+            <div className="mt-8 space-y-4">
+              <div className="flex items-start space-x-3">
+                <MapPinIcon className="w-6 h-6 text-[#06B7DB] stroke-2" />
+                <div>
+                  <p className="text-gray-700 dark:text-gray-300">451 Health Sciences Dr.,</p>
+                  <p className="text-gray-700 dark:text-gray-300">Davis, CA 95616</p>
+                </div>
+              </div>
+              
+              <div className="flex items-center space-x-3">
+                <PhoneIcon className="w-6 h-6 text-[#06B7DB] stroke-2" />
+                <p className="text-gray-700 dark:text-gray-300">(530) 754-9654</p>
+              </div>
+            </div>
           </div>
           <Card className="bg-white dark:bg-gray-900 shadow-lg-top w-full max-w-4xl">
             <CardBody className="p-6 md:p-8">
@@ -189,7 +212,12 @@ const ContactUs = () => {
       
       <Modal isOpen={isOpen} onClose={onClose}>
         <ModalContent>
-          <ModalHeader>{modalContent.title}</ModalHeader>
+          <ModalHeader className="flex items-center gap-2">
+            {modalContent.type === "success" && <FaCheckCircle className="text-green-500 text-xl" />}
+            {modalContent.type === "error" && <FaExclamationTriangle className="text-red-500 text-xl" />}
+            {modalContent.type === "loading" && <FaPaperPlane className="text-[#06B7DB] text-xl animate-bounce" />}
+            {modalContent.title}
+          </ModalHeader>
           <ModalBody className="py-6">
             {isLoading ? (
               <div className="flex flex-col items-center gap-4">
@@ -197,7 +225,9 @@ const ContactUs = () => {
                 <p>{modalContent.message}</p>
               </div>
             ) : (
-              <p>{modalContent.message}</p>
+              <div className="flex items-center gap-4">
+                <p>{modalContent.message}</p>
+              </div>
             )}
           </ModalBody>
         </ModalContent>
